split up stack trace generation into multiple files
